Make all providers into interfaces
This will allow post generation changes to generated SQL
