+++ src/main/java/org/mybatis/dynamic/sql/delete/render/DeleteRenderer.java
@@ -1,5 +1,5 @@
 /**
- *    Copyright 2016-2017 the original author or authors.
+ *    Copyright 2016-2018 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -32,7 +32,7 @@ public class DeleteRenderer {
     }
     
     public DeleteStatementProvider render() {
-        return DeleteStatementProvider.withTableName(deleteModel.table().name())
+        return DefaultDeleteStatementProvider.withTableName(deleteModel.table().name())
                 .withWhereClause(deleteModel.whereModel().map(this::renderWhereClause))
                 .build();
     }
