+++ src/main/java/org/mybatis/dynamic/sql/insert/render/InsertSelectRenderer.java
@@ -1,5 +1,5 @@
 /**
- *    Copyright 2016-2017 the original author or authors.
+ *    Copyright 2016-2018 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -38,7 +38,7 @@ public class InsertSelectRenderer {
     public InsertSelectStatementProvider render() {
         SelectStatementProvider selectStatement = model.selectModel().render(renderingStrategy);
         
-        return InsertSelectStatementProvider.withTableName(model.table().name())
+        return DefaultInsertSelectStatementProvider.withTableName(model.table().name())
                 .withColumnsPhrase(calculateColumnsPhrase())
                 .withSelectStatement(selectStatement.getSelectStatement())
                 .withParameters(selectStatement.getParameters())
