+++ src/main/java/org/mybatis/dynamic/sql/insert/render/InsertSelectStatementProvider.java
@@ -1,5 +1,5 @@
 /**
- *    Copyright 2016-2017 the original author or authors.
+ *    Copyright 2016-2018 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -15,69 +15,10 @@
  */
 package org.mybatis.dynamic.sql.insert.render;
 
-import static org.mybatis.dynamic.sql.util.StringUtilities.spaceBefore;
-
-import java.util.HashMap;
 import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
 
-public class InsertSelectStatementProvider {
-    private String tableName;
-    private Optional<String> columnsPhrase;
-    private String selectStatement;
-    private Map<String, Object> parameters;
-    
-    private InsertSelectStatementProvider(Builder builder) {
-        tableName = Objects.requireNonNull(builder.tableName);
-        columnsPhrase = Objects.requireNonNull(builder.columnsPhrase);
-        selectStatement = Objects.requireNonNull(builder.selectStatement);
-        parameters = Objects.requireNonNull(builder.parameters);
-    }
-    
-    public String getInsertStatement() {
-        return "insert into" //$NON-NLS-1$
-                + spaceBefore(tableName)
-                + spaceBefore(columnsPhrase)
-                + spaceBefore(selectStatement);
-    }
-    
-    public Map<String, Object> getParameters() {
-        return parameters;
-    }
+public interface InsertSelectStatementProvider {
+    Map<String, Object> getParameters();
 
-    public static Builder withTableName(String tableName) {
-        return new Builder().withTableName(tableName);
-    }
-    
-    public static class Builder {
-        private String tableName;
-        private Optional<String> columnsPhrase;
-        private String selectStatement;
-        private Map<String, Object> parameters = new HashMap<>();
-        
-        public Builder withTableName(String tableName) {
-            this.tableName = tableName;
-            return this;
-        }
-
-        public Builder withColumnsPhrase(Optional<String> columnsPhrase) {
-            this.columnsPhrase = columnsPhrase;
-            return this;
-        }
-        
-        public Builder withSelectStatement(String selectStatement) {
-            this.selectStatement = selectStatement;
-            return this;
-        }
-        
-        public Builder withParameters(Map<String, Object> parameters) {
-            this.parameters.putAll(parameters);
-            return this;
-        }
-        
-        public InsertSelectStatementProvider build() {
-            return new InsertSelectStatementProvider(this);
-        }
-    }
+    String getInsertStatement();
 }
